Cleaning up.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@2387 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ Testprog/TreeFilter/tfTest/main.cc
@@ -33,7 +33,7 @@ namespace v
 ************************************************************************/
 enum	{c_Sigma, c_Normalization, c_Saturation};
 
-static int	range[][3] = {{1, 255, 2}, {1, 255, 1}};
+static float	range[][3] = {{1, 255, 2}, {1, 255, 1}};
 static CmdDef	Cmds[] =
 {
     {C_Slider, c_Sigma,		11, "Sigma:",		range[0], CA_None,
